Extract site-builder.js
